+++ app/packages/state/src/hooks/schema/useSetSelectedFieldsStage.ts
@@ -1,0 +1,48 @@
+import * as fos from "@fiftyone/state";
+import * as foq from "@fiftyone/relay";
+import { useContext } from "react";
+import { useRecoilCallback, useRecoilValue } from "recoil";
+import { useMutation } from "react-relay";
+
+/**
+ *
+ * @returns a callback to set the selectedFields stage
+ */
+export default function useSetSelectedFieldsStage() {
+  const datasetName = useRecoilValue(fos.datasetName);
+  const router = useContext(fos.RouterContext);
+  const [setView] = useMutation<foq.setViewMutation>(foq.setView);
+
+  return useRecoilCallback(
+    ({ snapshot, set }) =>
+      async (value) => {
+        if (!datasetName) {
+          return;
+        }
+
+        set(fos.selectedFieldsStageState, value);
+
+        // router is loaded only in OSS
+        if (router.loaded) return;
+        const view = await snapshot.getPromise(fos.view);
+        const subscription = await snapshot.getPromise(fos.stateSubscription);
+        setView({
+          variables: {
+            view: value ? [...view, value] : view,
+            datasetName: datasetName,
+            form: {},
+            subscription,
+          },
+          onCompleted: ({ setView: { dataset } }) => {
+            // in an embedded context, we update the dataset schema through the
+            // state proxy
+            set(fos.stateProxy, (current) => ({
+              ...(current || {}),
+              dataset,
+            }));
+          },
+        });
+      },
+    [setView, router, datasetName]
+  );
+}
